Update GuiApi so that TestDrawingWithTwoScripts passes. This fixes the bug where if two lua scripts draw stuff on the same frame only one script's drawings are visible. Replaces a bunch of obscure ThisIsTheLuaHack and locking/unlocking code with a simple pair of methods.

+++ src/BizHawk.Tests/Client.Common/lua/TestLuaDrawing.cs
@@ -71,7 +71,7 @@ namespace BizHawk.Tests.Client.Common.Lua
 			var buffer = displayManager.RenderOffscreenLua(vp);
 
 			Assert.AreEqual(buffer.GetPixel(2, 2), Color.Red.ToArgb());
-			Assert.AreEqual(buffer.GetPixel(2, 4), 0xff00ff00);
+			Assert.AreEqual((uint)buffer.GetPixel(2, 4), 0xff00ff00);
 		}
 
 	}
